do not delete hero image if it does not exist

+++ src/Repositories/PageRepository.php
@@ -70,7 +70,9 @@ class PageRepository extends CmsRepository
         $payload = $this->parseBlocks($payload, 'pages');
 
         if (isset($payload['hero_image'])) {
-            app(FileService::class)->delete($page->hero_image);
+            if ($page->hero_image) {
+                app(FileService::class)->delete($page->hero_image);
+            }
             $file = request()->file('hero_image');
             $path = app(FileService::class)->saveFile($file, 'public/images', [], true);
             $payload['hero_image'] = $path['name'];
